requirements.txt
fastapi
langchain
langgraph
langchain-core
langchain-community
langchain-nvidia-ai-endpoints
python-dotenv
requests
beautifulsoup4
wikipedia
duckduckgo-search
uvicorn
pydantic
